server actions splitted and created new folder for them
